refactor: DataChannelTopic lives in types, not the livekit manager
The topic enum was declared inside the LiveKit manager, so the shared
StreamingManager type and the AgentManager type could only reach it by
importing from a service - a types -> service edge, and a cycle waiting
to happen once the primitive wanted the type. It moves to
src/types/stream next to StreamEvents and the package root keeps
exporting it (via `export * from './types'`), so the public surface is
unchanged.

With the enum reachable, the send primitive is typed
`topic: DataChannelTopic` on the shared type and both implementations,
and so is AgentManager.sendDataMessage. V1's interrupt now passes
DataChannelTopic.Interrupt instead of a StreamEvents member: WebRTC
ignores the topic entirely (the payload's `type` identifies the
interrupt), which the call site now says out loud.
