+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from './errors';
 export * from './services/agent-manager';
-export { DataChannelTopic } from './services/streaming-manager/livekit-manager';
 export * from './types';
 export { parseMessageParts } from './utils/content-parser';
 export { isAwaitingTool } from './utils/tool-calls';
