+++ src/services/streaming-manager/edge-cases.test.ts
@@ -8,6 +8,7 @@ import {
     AgentActivityState,
     ConnectionState,
     CreateStreamOptions,
+    DataChannelTopic,
     StreamType,
     StreamingManagerOptions,
 } from '../../types/index';
@@ -68,7 +69,7 @@ describe('Streaming Manager Edge Cases', () => {
             const mockDC = mockPC.createDataChannel.mock.results[0].value;
 
             mockDC.readyState = 'connecting';
-            manager.sendDataChannelMessage('did.test', 'test');
+            manager.sendDataChannelMessage(DataChannelTopic.Interrupt, 'test');
             expect(options.callbacks.onError).toHaveBeenCalled();
 
             mockDC.readyState = 'open';
@@ -232,15 +233,15 @@ describe('Streaming Manager Edge Cases', () => {
             const mockDC = mockPC.createDataChannel.mock.results[0].value;
 
             mockDC.readyState = 'connecting';
-            manager.sendDataChannelMessage('did.test', 'test1');
+            manager.sendDataChannelMessage(DataChannelTopic.Interrupt, 'test1');
             expect(options.callbacks.onError).toHaveBeenCalled();
 
             mockDC.readyState = 'closing';
-            manager.sendDataChannelMessage('did.test', 'test2');
+            manager.sendDataChannelMessage(DataChannelTopic.Interrupt, 'test2');
             expect(options.callbacks.onError).toHaveBeenCalled();
 
             mockDC.readyState = 'closed';
-            manager.sendDataChannelMessage('did.test', 'test3');
+            manager.sendDataChannelMessage(DataChannelTopic.Interrupt, 'test3');
             expect(options.callbacks.onError).toHaveBeenCalled();
         });
 
@@ -538,7 +539,7 @@ describe('Streaming Manager Edge Cases', () => {
             mockPC.iceConnectionState = 'connected';
             mockPC.oniceconnectionstatechange();
 
-            manager.sendDataChannelMessage('did.test', 'test-message');
+            manager.sendDataChannelMessage(DataChannelTopic.Interrupt, 'test-message');
 
             await manager.disconnect();
